split off WorldPalette module
